+++ demisto_sdk/commands/validate/tests/validators_test.py
@@ -1783,3 +1783,17 @@ def test_job_unexpected_field_values_in_non_feed_job(repo, capsys,
     stdout = capsys.readouterr().out
     assert "Job must either have non-empty selectedFeeds OR have isAllFields set to true when isFeed is set to true" \
            in stdout
+
+@pytest.mark.parametrize('file_set,expected_output', (({'file_path'}, "[ST113] - file file_path was deleted from git, please restore the file."),
+                                                    ({}, "no deleted files were found."))
+                        )
+def test_validate_deleted_files(capsys, file_set, expected_output):
+
+    validate_manager = ValidateManager(check_is_unskipped=False, skip_conf_json=True)
+
+    validate_manager.validate_deleted_files(file_set)
+
+    stdout = capsys.readouterr().out
+
+    assert expected_output in stdout
+