+++ demisto_sdk/tests/integration_tests/validate_integration_test.py
@@ -642,6 +642,10 @@ class TestDeprecatedIntegration:
         mocker.patch.object(GitUtil, '__init__', return_value=None)
         mocker.patch.object(GitUtil, 'get_current_working_branch', return_value='MyBranch')
 
+        mocker.patch.object(ValidateManager, 'get_deleted_files_from_git', return_value={})                                                        
+        mocker.patch.object(ValidateManager, 'validate_deleted_files', return_value=True)
+        mocker.patch.object(GitUtil, 'deleted_files', return_value={})
+
         with ChangeCWD(pack.repo_path):
             runner = CliRunner(mix_stderr=False)
             result = runner.invoke(main, [VALIDATE_CMD, '-g', '--no-docker-checks',
@@ -709,6 +713,10 @@ class TestDeprecatedIntegration:
                                                                                          set(), set()))
         mocker.patch.object(GitUtil, '__init__', return_value=None)
         mocker.patch.object(GitUtil, 'get_current_working_branch', return_value='MyBranch')
+
+        mocker.patch.object(ValidateManager, 'get_deleted_files_from_git', return_value={})                                                        
+        mocker.patch.object(ValidateManager, 'validate_deleted_files', return_value=True)
+        mocker.patch.object(GitUtil, 'deleted_files', return_value={})
 
         with ChangeCWD(pack.repo_path):
             runner = CliRunner(mix_stderr=False)
@@ -2316,6 +2324,10 @@ class TestPlaybookValidateDeprecated:
         mocker.patch.object(GitUtil, '__init__', return_value=None)
         mocker.patch.object(GitUtil, 'get_current_working_branch', return_value='MyBranch')
 
+        mocker.patch.object(ValidateManager, 'get_deleted_files_from_git', return_value={})                                                        
+        mocker.patch.object(ValidateManager, 'validate_deleted_files', return_value=True)
+        mocker.patch.object(GitUtil, 'deleted_files', return_value={})
+
         with ChangeCWD(pack.repo_path):
             runner = CliRunner(mix_stderr=False)
             result = runner.invoke(main,
@@ -2382,6 +2394,10 @@ class TestPlaybookValidateDeprecated:
                                                                                          set(), set()))
         mocker.patch.object(GitUtil, '__init__', return_value=None)
         mocker.patch.object(GitUtil, 'get_current_working_branch', return_value='MyBranch')
+
+        mocker.patch.object(ValidateManager, 'get_deleted_files_from_git', return_value={})                                                        
+        mocker.patch.object(ValidateManager, 'validate_deleted_files', return_value=True)
+        mocker.patch.object(GitUtil, 'deleted_files', return_value={})
 
         with ChangeCWD(pack.repo_path):
             runner = CliRunner(mix_stderr=False)
@@ -2647,6 +2663,10 @@ class TestScriptDeprecatedValidation:
         mocker.patch.object(GitUtil, '__init__', return_value=None)
         mocker.patch.object(GitUtil, 'get_current_working_branch', return_value='MyBranch')
 
+        mocker.patch.object(ValidateManager, 'get_deleted_files_from_git', return_value={})                                                        
+        mocker.patch.object(ValidateManager, 'validate_deleted_files', return_value=True)
+        mocker.patch.object(GitUtil, 'deleted_files', return_value={})
+
         with ChangeCWD(pack.repo_path):
             runner = CliRunner(mix_stderr=False)
             result = runner.invoke(main,
@@ -2711,6 +2731,10 @@ class TestScriptDeprecatedValidation:
                                                                                          set(), set()))
         mocker.patch.object(GitUtil, '__init__', return_value=None)
         mocker.patch.object(GitUtil, 'get_current_working_branch', return_value='MyBranch')
+
+        mocker.patch.object(ValidateManager, 'get_deleted_files_from_git', return_value={})                                                        
+        mocker.patch.object(ValidateManager, 'validate_deleted_files', return_value=True)
+        mocker.patch.object(GitUtil, 'deleted_files', return_value={})
 
         with ChangeCWD(pack.repo_path):
             runner = CliRunner(mix_stderr=False)
@@ -3047,6 +3071,10 @@ class TestValidationUsingGit:
         mocker.patch.object(GitUtil, '__init__', return_value=None)
         mocker.patch.object(GitUtil, 'get_current_working_branch', return_value='MyBranch')
 
+        mocker.patch.object(ValidateManager, 'get_deleted_files_from_git', return_value={})                                                        
+        mocker.patch.object(ValidateManager, 'validate_deleted_files', return_value=True)
+        mocker.patch.object(GitUtil, 'deleted_files', return_value={})
+
         with ChangeCWD(repo.path):
             runner = CliRunner(mix_stderr=False)
             result = runner.invoke(main, [VALIDATE_CMD, '-g', '--no-docker-checks', '--no-conf-json',
@@ -3103,6 +3131,10 @@ class TestValidationUsingGit:
         mocker.patch.object(GitUtil, '__init__', return_value=None)
         mocker.patch.object(GitUtil, 'get_current_working_branch', return_value='MyBranch')
 
+        mocker.patch.object(ValidateManager, 'get_deleted_files_from_git', return_value={})                                                        
+        mocker.patch.object(ValidateManager, 'validate_deleted_files', return_value=True)
+        mocker.patch.object(GitUtil, 'deleted_files', return_value={})
+
         with ChangeCWD(repo.path):
             runner = CliRunner(mix_stderr=False)
             result = runner.invoke(main, [VALIDATE_CMD, '-g', '--no-docker-checks', '--no-conf-json',
@@ -3151,6 +3183,10 @@ class TestValidationUsingGit:
         mocker.patch.object(GitUtil, '__init__', return_value=None)
         mocker.patch.object(GitUtil, 'get_current_working_branch', return_value='MyBranch')
 
+        mocker.patch.object(ValidateManager, 'get_deleted_files_from_git', return_value={})                                                        
+        mocker.patch.object(ValidateManager, 'validate_deleted_files', return_value=True)
+        mocker.patch.object(GitUtil, 'deleted_files', return_value={})
+
         with ChangeCWD(repo.path):
             runner = CliRunner(mix_stderr=False)
             result = runner.invoke(main, [VALIDATE_CMD, '-g', '--no-docker-checks', '--no-conf-json',
@@ -3191,6 +3227,10 @@ class TestValidationUsingGit:
                                                                                          set(), set()))
         mocker.patch.object(GitUtil, '__init__', return_value=None)
         mocker.patch.object(GitUtil, 'get_current_working_branch', return_value='MyBranch')
+
+        mocker.patch.object(ValidateManager, 'get_deleted_files_from_git', return_value={})                                                        
+        mocker.patch.object(ValidateManager, 'validate_deleted_files', return_value=True)
+        mocker.patch.object(GitUtil, 'deleted_files', return_value={})
 
         with ChangeCWD(repo.path):
             runner = CliRunner(mix_stderr=False)
@@ -3274,6 +3314,10 @@ class TestValidationUsingGit:
         mocker.patch.object(GitUtil, '__init__', return_value=None)
         mocker.patch.object(GitUtil, 'get_current_working_branch', return_value='MyBranch')
 
+        mocker.patch.object(ValidateManager, 'get_deleted_files_from_git', return_value={})                                                        
+        mocker.patch.object(ValidateManager, 'validate_deleted_files', return_value=True)
+        mocker.patch.object(GitUtil, 'deleted_files', return_value={})
+
         with ChangeCWD(repo.path):
             runner = CliRunner(mix_stderr=False)
             result = runner.invoke(main, [VALIDATE_CMD, '-g', '--no-docker-checks', '--no-conf-json',
@@ -3314,6 +3358,10 @@ class TestValidationUsingGit:
                                                                                          set(), set()))
         mocker.patch.object(GitUtil, '__init__', return_value=None)
         mocker.patch.object(GitUtil, 'get_current_working_branch', return_value='MyBranch')
+
+        mocker.patch.object(ValidateManager, 'get_deleted_files_from_git', return_value={})                                                        
+        mocker.patch.object(ValidateManager, 'validate_deleted_files', return_value=True)
+        mocker.patch.object(GitUtil, 'deleted_files', return_value={})
 
         with ChangeCWD(repo.path):
             runner = CliRunner(mix_stderr=False)
@@ -3362,6 +3410,10 @@ class TestValidationUsingGit:
                                                                                          set(), set()))
         mocker.patch.object(GitUtil, '__init__', return_value=None)
         mocker.patch.object(GitUtil, 'get_current_working_branch', return_value='MyBranch')
+
+        mocker.patch.object(ValidateManager, 'get_deleted_files_from_git', return_value={})                                                        
+        mocker.patch.object(ValidateManager, 'validate_deleted_files', return_value=True)
+        mocker.patch.object(GitUtil, 'deleted_files', return_value={})
 
         with ChangeCWD(repo.path):
             runner = CliRunner(mix_stderr=False)